requirements.txt
bcrypt==3.2.2
click==8.1.3
uvicorn==0.17.6
starlette==0.20.1
pyjwt==2.4.0